% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{denormalize}
\alias{denormalize}
\title{denomalize data}
\usage{
denormalize(normalized, min, max)
}
\arguments{
\item{normalized}{the normalized data frame/matrix}

\item{min, max}{the min and max of each variable in the data. This can be otained by retrieving the attribute values of 
the normalized data, i.e. output of \code{normalize}}
}
\description{
tranform normalized data back to original scale
}
