% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmerBoost.R
\name{glmerBoost}
\alias{glmerBoost}
\alias{glmerLogitBoost}
\title{Boosted generalized mixed-effect regression models (glmer).}
\usage{
glmerBoost(form, dat, lme.family = binomial, max.iter = 100,
  verbose = FALSE, coeflearn = "Breiman", ...)

glmerBoost(form, dat, lme.family = binomial, max.iter = 100,
  verbose = FALSE, coeflearn = "Breiman", ...)

glmerLogitBoost(form, dat, max.iter = 100, verbose = FALSE, ...)

glmerLogitBoost(form, dat, max.iter = 100, verbose = FALSE, ...)
}
\arguments{
\item{form}{A formula describing fixed and  random effects (as in lme4)}

\item{dat}{Matrix/data frame}

\item{lme.family}{family for glmer}

\item{max.iter}{maximum iteration}

\item{verbose}{print iteration outputs ?}

\item{coeflearn}{Adaboot.M1 learning method}

\item{\dots}{other arguments}
}
\value{
a list with items 
\item{tree.fit}{fitted lme4 models}
\item{fitted.probs}{fitted probabilites for final model}
\item{fitted.class}{fitted class labels for final model}
\item{train.perf}{various performance measures for final model on training set}
\item{threshold}{classification cut-off}
}
\description{
Boosted generalized mixed-effect regression models (glmer).
}
\examples{
\dontrun{
set.seed(12345)
mod <- glmerBoost(form, data)) 
}
\dontrun{
set.seed(12345)
mod <- glmerBoost(form, rand.form, data)) 
}
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}
