% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{LongiLagSplit}
\alias{LongiLagSplit}
\title{Train-Test  longitudinal data split}
\usage{
LongiLagSplit(dat, id, rhs.vars, resp.vars, order.vars = "time",
  lag = 1)
}
\description{
Split longitudinal data into training and test set, where the training set lags the 
test set by an amount "lag". e.g predict an outcome at lag=x hospital visits in advanced.  
There must be at least 2 repeated measurements for each unit
}
