% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEgbm.R
\name{MEgbm}
\alias{MEgbm}
\alias{MEgbmRules}
\title{Mixed Effect GBM}
\usage{
MEgbm(X, Y, groups = NULL, rand.vars = "1", para = NULL,
  lme.family = binomial, tol = 1e-05, max.iter = 100,
  include.RE = TRUE, verbose = FALSE, likelihoodCheck = TRUE, ...)

MEgbm(X, Y, groups = NULL, rand.vars = "1", para = NULL,
  lme.family = binomial, tol = 1e-05, max.iter = 100,
  include.RE = TRUE, verbose = FALSE, likelihoodCheck = TRUE, ...)

MEgbmRules(form, dat, groups = NULL, rand.vars = "1", para = NULL,
  tol = 1e-05, max.iter = 100, include.RE = FALSE, verbose = FALSE,
  maxdepth = 5, glmer.Control = glmerControl(optimizer = "bobyqa",
  check.nobs.vs.nRE = "ignore", check.nobs.vs.nlev = "ignore"), nAGQ = 0,
  likelihoodCheck = TRUE, K = 3, decay = 0.05, ...)

MEgbmRules(form, dat, groups = NULL, rand.vars = "1", para = NULL,
  tol = 1e-05, max.iter = 100, include.RE = FALSE, verbose = FALSE,
  maxdepth = 5, glmer.Control = glmerControl(optimizer = "bobyqa",
  check.nobs.vs.nRE = "ignore", check.nobs.vs.nlev = "ignore"), nAGQ = 0,
  likelihoodCheck = TRUE, K = 3, decay = 0.05, ...)
}
\arguments{
\item{X}{data.frame with predictors}

\item{Y}{binary response vector}

\item{groups}{character name of the column containing the group identifier}

\item{rand.vars}{random effect variables}

\item{para}{named list of gbm training parameters}

\item{lme.family}{glmer control}

\item{tol}{convergence tolerance}

\item{max.iter}{maximum number of iterations}

\item{include.RE}{(logical) to include random effect Zb as predictor in gbm?}

\item{verbose}{verbose for lme4}

\item{likelihoodCheck}{(logical) to use log likelihood of glmer to check for convergence?}

\item{\dots}{Further arguments passed to or from other methods.}

\item{gbm.dist}{gbm loss function}

\item{type}{of predictions of gbm to pass to lme4 as population estimates (these will be used as offset)}
}
\value{
An object of class MEgbm; a list with items 
\item{gbmfit}{fitted gbm model}
\item{glmer.fit}{fitted mixed effect logistic regression model}
\item{logLik}{log likelihood of mixed effect logistic regression} 
\item{random.effects}{random effect parameter estimates}
\item{boost.form}{gbm formula for fitted model}
\item{glmer.form}{lmer4 formula} 
\item{glmer.CI}{estimates of mixed effect logistic regression with 
    approximate confidence intervals on the logit scale. More accurate values 
    can be obtained by bootstrap}
\item{fitted.probs}{fitted probabilites for final model}
\item{fitted.class}{fitted class labels for final model}
\item{train.perf}{various performance measures for final model on training set}
\item{threshold}{classification cut-off}
}
\description{
Trains a Mixed Effect gradient boosted machine  
for longitudinal continuous, binary and count data. A rule based version or these methods 
using the \code{inTree} package is also implemented(see [1])
}
\examples{
\dontrun{
set.seed(12345)
mod <- MEgbm(form, rand.form, data)) 

}
}
\references{
Che Ngufor,  Holly Van Houten, Brian S. Caffo , Nilay D. Shah, Rozalina G. McCoy 
Mixed Effect Machine Learning: a framework for predicting longitudinal change in hemoglobin A1c,
in Journal of Biomedical Informatics, 2018
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}
