% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEglm.R
\name{MEglm}
\alias{MEglm}
\title{Mixed Effect GLM}
\usage{
MEglm(form, data, control, nAGQ = 0)

MEglm(form, data, control, nAGQ = 0)
}
\arguments{
\item{form}{formula}

\item{data}{data.frame with predictors}

\item{nAGQ}{see \code{glmer}}

\item{Control}{glmer control}
}
\description{
Trains a Mixed Effect generalized mixed effect models. This function just calls glmer from the Lme4 package.
}
\examples{
\dontrun{
set.seed(12345)
mod <- MEglm(form, data= dat.trn, family=binomial) 

}
}
