% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEglmTree.R
\name{MEglmTree}
\alias{MEglmTree}
\alias{MECTree}
\title{Trains a Mixed Effect Model Based Recursive partitioning and Mixed Effect 
 classification trees for longitudinal continuous, binary and count data. 
 These functions can be called directly and more efficiently by the functions \code{MEml} and \code{MEml.lag} 
#}
\usage{
MEglmTree(X, Y, part.vars, reg.vars = "1", rand.vars = "1",
  groups = NULL, initialRandomEffects = rep(0, nrow(X)),
  initialProbs = rep(0.5, nrow(X)), tol = 1e-05, max.iter = 100,
  include.RE = TRUE, verbose = FALSE, likelihoodCheck = TRUE,
  glmer.Control = glmerControl(optimizer = "bobyqa", check.nobs.vs.nRE =
  "ignore", check.nobs.vs.nlev = "ignore"), nAGQ = 0, alpha = 0.05,
  minsize = 20, maxdepth = 50, para = NULL, ...)

MEglmTree(X, Y, part.vars, reg.vars = "1", rand.vars = "1",
  groups = NULL, initialRandomEffects = rep(0, nrow(X)),
  initialProbs = rep(0.5, nrow(X)), tol = 1e-05, max.iter = 100,
  include.RE = TRUE, verbose = FALSE, likelihoodCheck = TRUE,
  glmer.Control = glmerControl(optimizer = "bobyqa", check.nobs.vs.nRE =
  "ignore", check.nobs.vs.nlev = "ignore"), nAGQ = 0, alpha = 0.05,
  minsize = 20, maxdepth = 50, para = NULL, ...)

MECTree(X, Y, con.tree = FALSE, rhs.vars, rand.vars = "1",
  groups = NULL, tol = 1e-05, max.iter = 100, verbose = FALSE,
  likelihoodCheck = TRUE, glmer.Control = glmerControl(optimizer =
  "bobyqa", check.nobs.vs.nRE = "ignore", check.nobs.vs.nlev = "ignore"),
  nAGQ = 0, cv = TRUE, cpmin = 1e-04, minsplit = 50,
  minbucket = 10, no.SE = 1, mincriterion = 0.975, maxdepth = 30,
  stump = FALSE, ...)

MECTree(X, Y, con.tree = FALSE, rhs.vars, rand.vars = "1",
  groups = NULL, tol = 1e-05, max.iter = 100, verbose = FALSE,
  likelihoodCheck = TRUE, glmer.Control = glmerControl(optimizer =
  "bobyqa", check.nobs.vs.nRE = "ignore", check.nobs.vs.nlev = "ignore"),
  nAGQ = 0, cv = TRUE, cpmin = 1e-04, minsplit = 50,
  minbucket = 10, no.SE = 1, mincriterion = 0.975, maxdepth = 30,
  stump = FALSE, ...)
}
\arguments{
\item{X}{data.frame with predictors}

\item{Y}{binary response vector}

\item{part.vars, rhs.vars}{partitioning variables for MOB and predictors}

\item{reg.vars}{regressors for MOB}

\item{rand.vars}{random effect variables}

\item{groups}{character name of the column containing the group identifier}

\item{initialRandomEffects}{[0] a vector of initial values for random effects}

\item{initialProbs}{[0.5]  a vector of initial conditional probabilites of success}

\item{tol}{convergence tolerance}

\item{max.iter}{maximum number of iteration for the EM algorithm}

\item{include.RE}{include random effects in glmtree model part?}

\item{verbose}{logical for printing intermediate trees results}

\item{likelihoodCheck}{logical: should the likelihood of random effect model be used to check for convergence?}

\item{glmer.Control}{same as \code{\link[lme4]{glmerControl}} - glmer controls, 
default to glmerControl(optimizer = "bobyqa")]}

\item{nAGQ}{as in \code{\link[lme4]{glmer}}, default to 10 
with maximum likelihood or restricted maximum likelihood}

\item{alpha}{stability parameter in glmtree}

\item{minsize, maxdepth, minsplit, minbucket, mincriterion, stump}{see rpart}

\item{\dots}{Further arguments passed to or from other methods.}

\item{con.tree}{do conditional inference trees instead of rpart?}

\item{cv}{[TRUE] - Should cross-validation be used?}

\item{cpmin}{[0.0MEBoostedTree001] - complexity parameter used in building a tree before cross-validation}

\item{no.SE}{[0] - number of standard errors used in pruning (0 if unused)}
}
\value{
An object of class \code{MECTree}; a list with items 
\item{tree.fit}{fitted classification trees model}
\item{glmer.fit}{fitted mixed effect logistic regression model}
\item{logLik}{log likelihood of mixed effect logistic regression} 
\item{random.effects}{random effect parameter estimates}
\item{form}{modified formula for fitted classification trees model}
\item{rand.form}{modified formula for random effects} 
\item{glmer.Control}{glmer controls}
\item{tree.control}{rpart controls}
\item{glmer.CI}{estimates of mixed effect logistic regression with 
    approximate confidence intervals on the logit scale. More accurate values 
    can be obtained by bootstrap}
\item{fitted.probs}{fitted probabilites for final model}
\item{fitted.class}{fitted class labels for final model}
\item{train.perf}{various performance measures for final model on training set}
\item{threshold}{classification cut-off}
}
\description{
Trains a Mixed Effect Model Based Recursive partitioning and Mixed Effect 
 classification trees for longitudinal continuous, binary and count data. 
 These functions can be called directly and more efficiently by the functions \code{MEml} and \code{MEml.lag} 
#

#
}
\examples{
\dontrun{
set.seed(12345)
mod <- MEMOBTree(form, rand.form, data)) 
#
}
\dontrun{
set.seed(12345)
mod <-MECTree(fix.form, rand.form, data)) 

}
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}
