% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEml.R
\name{MEml}
\alias{MEml}
\alias{MEml_lag}
\alias{MEml2}
\title{Calls the MEml models: MEgbm,  MErf, and MEglm. 
The training and test data can be split into lagged training and testing as described in [1]}
\usage{
MEml_lag(lag = NULL, classifier, dat, id, rhs.vars, resp.vars,
  order.vars, rand.vars = NULL, reg.vars = NULL, part.vars = NULL,
  para, max.iter = 10, seed = 1, return.model = TRUE)

MEml(classifier, dat.trn, dat.tst, id, rhs.vars, resp.vars,
  rand.vars = NULL, reg.vars = NULL, part.vars = NULL, para,
  max.iter = 10, seed = 1, return.model = FALSE, ...)

MEml2(method, data, id, resp.vars, rhs.vars, rand.vars = NULL, para, ...)
}
\arguments{
\item{lag}{time lag between predictors and outcome: e.g if lag = 1, then we use predictors in current
vist to predict outcome in the next visit.}

\item{dat}{data frame with predictors and outcome}

\item{id}{character name of the column containing the group identifier}

\item{rhs.vars}{caracter vector of predictors}

\item{order.vars}{order variables (usually time variable)}

\item{rand.vars}{random effect variables}

\item{reg.vars}{reg.vars regressors for MOB}

\item{part.vars}{partitioning variables for MOB and predictors}

\item{para}{named list of gbm training parameters}

\item{max.iter}{maximum number of iterations}

\item{return.model}{should the train model be return. Otherwise the return values is only the performance metrics}

\item{method}{string name of model. See names(\code{MEml_train}()) for list of possible models.}
}
\value{
The train MEml model and performance matrics (as data frame) if return.model = TRUE
}
\description{
Calls the MEml models: MEgbm,  MErf, and MEglm. 
The training and test data can be split into lagged training and testing as described in [1]
}
\details{
\enumerate{
 \item \code{MEml_lag} Takes the full data set and calls \code{LongiLagSplit} to split data into lagged 
 training and testing.  \code{MEml_lag} also trains the MOB and CTree models (see [1]).  
 \item  \code{MEml} is the same as \code{MEml_lag}, except that you pass in the training and test set. So you can 
 call \code{LongiLagSlit} and pass the derived training and test sets to \code{MEml}.  
 \item  \code{MEml2} is the same as \code{MEml}, except that you don't pass in the test set. 
 Also, it is currently implemented only for the GLMER, MEgbm, MErf, and GBM models.  
}
}
\references{
Che Ngufor,  Holly Van Houten, Brian S. Caffo , Nilay D. Shah, Rozalina G. McCoy 
Mixed Effect Machine Learning: a framework for predicting longitudinal change in hemoglobin A1c, 
in Journal of Biomedical Informatics, 2018 
#
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}
