% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MErf.R
\name{MErf}
\alias{MErf}
\title{Mixed Effect random forest}
\usage{
MErf(form, dat, groups = NULL, family, rand.vars = "1", para = NULL,
  tol = 1e-05, max.iter = 100, include.RE = FALSE, verbose = FALSE,
  maxdepth = 5, glmer.Control = glmerControl(optimizer = "bobyqa"),
  nAGQ = 0, likelihoodCheck = TRUE, K = 3, decay = 0.05, ...)
}
\arguments{
\item{form}{formula}

\item{dat}{data.frame with predictors}

\item{groups}{character name of the column containing the group identifier}

\item{family}{a GLM family: continuous data set family = "gaussian", binary data set 
family = "binomial" see \code{glm} and \code{family}}

\item{rand.vars}{random effect variables}

\item{para}{named list of gbm training parameters}

\item{tol}{convergence tolerance}

\item{max.iter}{maximum number of iterations}

\item{include.RE}{(logical) to include random effect Zb as predictor in gbm?}

\item{verbose}{verbose for lme4}

\item{glmer.Control}{\code{glmer} or \code{lmer} control}

\item{likelihoodCheck}{(logical) to use log likelihood of glmer to check for convergence?}

\item{\dots}{Further arguments passed to or from other methods.}

\item{type}{of predictions of gbm to pass to lme4 as population estimates (these will be used as offset)}
}
\value{
An object of class MEgbm; a list with items 
\item{rf.fit}{fitted random forest model}
\item{glmer.fit}{fitted mixed effect logistic regression model}
\item{logLik}{log likelihood of mixed effect logistic regression} 
\item{random.effects}{random effect parameter estimates}
\item{glmer.form}{lmer4 formula} 
\item{glmer.CI}{estimates of mixed effect logistic regression with 
    approximate confidence intervals on the logit scale. More accurate values 
    can be obtained by bootstrap}
\item{threshold}{classification cut-off}
\item{predRules}{fitted rules}
\item{Y.star}{fitted "transform" outcome. This is the same as the predicted outcome for binary data}
}
\description{
Trains a Mixed Effect random forest 
for longitudinal continuous  and binary 
data. A rule based version or these methods 
using the \code{inTree} package is also implemented(see [1])
}
\references{
Che Ngufor,  Holly Van Houten, Brian S. Caffo , Nilay D. Shah, Rozalina G. McCoy 
Mixed Effect Machine Learning: a framework for predicting longitudinal change in hemoglobin A1c,
in Journal of Biomedical Informatics, 2018
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}
