% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{normalize}
\alias{normalize}
\title{normalize data}
\usage{
normalize(x)
}
\arguments{
\item{x}{matrix or data frame}
}
\value{
nomalized matrix/data frame with min and max of each variable as attributes
}
\description{
Normalize matrix or data frame to the min-max range of the variables.
}
