% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Performance.measures}
\alias{Performance.measures}
\title{Performance measures}
\usage{
Performance.measures(pred, obs, threshold = NULL, prevalence = NULL)
}
\arguments{
\item{pred}{predicted probabilities}

\item{obs}{true class}

\item{threshold}{class probability threshold (default to NULL)}

\item{prevalence}{class prevalence in the population (default to NULL)}
}
\description{
Compute accuracy, AUC, sensitivity, specificity, positive predictive values, etc.
}
