% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MEglmTree.R
\name{Plot.MEglmTree}
\alias{Plot.MEglmTree}
\alias{plot.MEglmTree}
\alias{plot.MECTree}
\alias{print.MECTree}
\alias{heat.tree}
\alias{.plot_node}
\alias{NodeBoxplot}
\alias{NodeBarplot}
\title{Plot method for MEglmTree object
Take a MEglmTree and plot the barplot of fitted classes or boxplot of 
fitted probability}
\usage{
\method{plot}{MEglmTree}(x, main = NULL,
  terminal_panel = node_terminal, tp_args = list(),
  inner_panel = node_inner, ip_args = list(),
  edge_panel = edge_simple, ep_args = list(), drop_terminal = FALSE,
  tnex = 1, newpage = TRUE, pop = TRUE, gp = gpar(), ...)

\method{plot}{MECTree}(x, ...)

\method{print}{MECTree}(x, ...)

heat.tree(tree, low.is.green = FALSE, ...)

.plot_node(node, xlim, ylim, nx, ny, terminal_panel, inner_panel,
  edge_panel, tnex = 2, drop_terminal = TRUE, debug = FALSE)

NodeBoxplot(obj, fitted = TRUE, col = "black", fill = "lightgray",
  width = 0.5, yscale = NULL, ylines = 3, cex = 0.5, id = TRUE,
  mainlab = NULL, gp = gpar(), ...)

NodeBarplot(obj, fitted = TRUE, col = "black", fill = NULL,
  beside = NULL, ymax = NULL, ylines = NULL, widths = 1,
  gap = NULL, reverse = NULL, id = TRUE, mainlab = NULL,
  gp = gpar(), ...)
}
\arguments{
\item{x}{fited MEglmTree or MECtree object.}

\item{main, ep_args}{see partykid}

\item{terminal_panel}{see partykit}

\item{tp_args, ip_args, xlim, ylim}{see partykit}

\item{inner_panel}{see partykit}

\item{edge_panel}{see partykit}

\item{drop_terminal}{see partykit}

\item{tnex}{see partykit}

\item{newpage, pop, gp, tree, low.is.green}{see partykit}

\item{\dots}{Further arguments passed to partykit}

\item{node}{see partykit}

\item{nx, ny}{see partykit}

\item{debug}{see partykit}
}
\description{
Plot method for MEglmTree object
Take a MEglmTree and plot the barplot of fitted classes or boxplot of 
fitted probability
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
