% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.glmerBoost.R
\name{predict.glmerBoost}
\alias{predict.glmerBoost}
\alias{predict.glmerLogitBoost}
\title{Make predictions using a fitted MECtree model}
\usage{
predict.glmerBoost(object, newdata, type = c("prob", "class")[1], ...)

predict.glmerLogitBoost(object, newdata, type = c("prob", "class")[1],
  ...)
}
\arguments{
\item{object}{Fitted model from MECtree.}

\item{newdata}{A new input data frame.}

\item{type}{of prediction: "prop" for probabilities and "class" for 
class labels.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with items 
\item{prob}{predicted class probabilities}
\item{class}{predicted class memberships obtained by thresholding 
class probabilities at the prevalence rate of the positive class}
}
\description{
Make predictions using a fitted MECtree model
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
