% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.MEgbm}
\alias{predict.MEgbm}
\title{Predict MEgbm   
Make predictions for Mixed Effect gbm.}
\usage{
predict.MEgbm(object, newdata, type = c("response", "prob", "class")[1],
  allow.new.levels = FALSE, ...)
}
\arguments{
\item{object}{Fitted model.}

\item{newdata}{A new input data frame.}

\item{type}{character string - either "response", the default, or "prob" or "class", indicating the type of prediction object returned.}

\item{allow.new.levels}{specifry if new levels of factor variables in the test set should be allowed. Default to FALSE.}
}
\value{
A list with items 
\item{Y.star}{predicted transformed outcome, same as predicted outcome for guassian distribution}
\item{pred}{matrix with predicted class probabilites, if type = "prob"} 
\item{predRules}{predicted rules}
}
\description{
Predict MEgbm   
Make predictions for Mixed Effect gbm.
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
