% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.MEgbmRules}
\alias{predict.MEgbmRules}
\title{Predict MEgbm   
Make predictions for Mixed Effect GBM.}
\usage{
predict.MEgbmRules(object, newdata, type = c("prob", "class")[1],
  allow.new.levels = FALSE, ...)
}
\arguments{
\item{object}{Fitted model.}

\item{newdata}{A new input data frame.}

\item{type}{of prediction: "prop" for probabilities and "class" for class labels.}

\item{allow.new.levels}{specifry if new levels of factor variables in the test set should be allowed. Default to FALSE.}
}
\description{
Predict MEgbm   
Make predictions for Mixed Effect GBM.
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
