% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.MEglm}
\alias{predict.MEglm}
\title{Predict MEglm  
Make predictions for Mixed Effect logistic regression. This is just a wrapper for \code{predict.merMod} in lme4 .}
\usage{
predict.MEglm(object, newdata, type = c("prob", "class")[1],
  allow.new.levels = FALSE, ...)
}
\arguments{
\item{object}{Fitted model.}

\item{newdata}{A new input data frame.}

\item{type}{of prediction: "prop" for probabilities and "class" for class labels.}

\item{allow.new.levels}{specifry if new levels of factor variables in the test set should be allowed. Default to FALSE.}
}
\value{
A list with items 
\item{Y.star}{NULL}
\item{pred}{matrix with predicted class probabilites}
\item{predRules}{NULL}
}
\description{
Predict MEglm  
Make predictions for Mixed Effect logistic regression. This is just a wrapper for \code{predict.merMod} in lme4 .
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
