% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.MErf}
\alias{predict.MErf}
\title{Predict MErf   
Make predictions for Mixed Effect random forest.}
\usage{
predict.MErf(object, newdata, type = c("response", "prob", "class")[1],
  allow.new.levels = FALSE, ...)
}
\arguments{
\item{object}{Fitted model.}

\item{newdata}{A new input data frame.}

\item{type}{of prediction: "prop" for probabilities and "class" for class labels.}

\item{allow.new.levels}{specifry if new levels of factor variables in the test set should be allowed. Default to FALSE.}
}
\value{
A list with items 
\item{Y.star}{predicted transformed outcome}
\item{pred}{matrix with predicted class probabilites} 
\item{predRules}{predicted rules}
}
\description{
Predict MErf   
Make predictions for Mixed Effect random forest.
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
