% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.MEgbmRules}
\alias{predict.MErfRules}
\alias{predict.MEglm}
\title{Predict MEml: MEglm, MEgbm and MErf   
Make predictions using fitted MEml2 mode fpr MEgbm and MErf.}
\usage{
predict.MEgbmRules(object, newdata, type = c("prob", "class")[1],
  allow.new.levels = FALSE, ...)

predict.MErfRules(object, newdata, type = c("prob", "class")[1],
  allow.new.levels = FALSE, ...)

predict.MEglm(object, newdata, type = c("prob", "class")[1],
  allow.new.levels = FALSE, ...)
}
\arguments{
\item{object}{Fitted mode.}

\item{newdata}{A new input data frame.}

\item{type}{of prediction: "prop" for probabilities and "class" for class labels.}

\item{allow.new.levels}{specifry if new levels of factor variables in the test set should be allowed. Default to FALSE.}
}
\description{
Predict MEml: MEglm, MEgbm and MErf   
Make predictions using fitted MEml2 mode fpr MEgbm and MErf.
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
