% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train.R
\name{Train}
\alias{Train}
\alias{TrainModels}
\alias{Train.Test}
\alias{Train.Boot}
\alias{cv.gbm}
\title{Train, Validate and Test various models:   
Functions to train, select best hyper-paramters, select best classification thresholds 
 and evaluate final model on the test set  using a number of machine 
learning classification algorithms}
\usage{
TrainModels(...)

Train.Test(...)

Train.Boot(classifier, nboots = 2, XY.dat, resp.vars, rhs.vars,
  part.vars, reg.vars, para, parallel = TRUE, n.cores = 2,
  seed = 12345678)

cv.gbm(X, y, outcome = NULL, n.trees = seq(150, 500, by = 10),
  interaction.depth, n.minobsinnode = 5, shrinkage = 10^(seq(-3, -1,
  length = 10)), bag.fraction = 0.5, distribution = "bernoulli",
  foldid = NULL, nfolds = 10, seed = 220, verbose = FALSE)
}
\arguments{
\item{\dots}{further arguments passed to or from other methods.}

\item{classifier}{character list of classification models. See names(TrainAllModels()).}

\item{para}{named list of model parameters}

\item{X.trn, X.val, X.tst}{matrix of predictors}

\item{Y.trn, Y.val, Y.tst}{matrix of binary \{0,1\} response variable}

\item{varimp}{(logical) compute variable importance ?}

\item{opt.para}{(logical) tune parameters ?}

\item{return.model}{(logical) return trained model ?}
}
\value{
\code{TrainModels} returns a list of functions for training various algorithms. Currently 
7 machine learning algorithms ELR, GLM, GLMnet, RF, GBM, avNNET, and SVM. Others can be easily added. 
\code{Train.Validate.Test} returns a list of performance measures for each classifier traoned:
\item{model}{trained model}
\item{para}{named list of model hyper-paramters (tunned values if opt.para = TRUE)}
\item{run.time}{compute time} 
\item{varimp}{variable importance if classifier is GLM, GBM or RF}
\item{perf}{a list with two data.frames: val.perf and tst.perf containing performance measures 
for validation and test sets }
}
\description{
Train, Validate and Test various models:   
Functions to train, select best hyper-paramters, select best classification thresholds 
 and evaluate final model on the test set  using a number of machine 
learning classification algorithms
}
\author{
Che Ngufor <Ngufor.Che@mayo.edu>
}
