% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils}
\alias{Utils}
\alias{reOnly}
\alias{LongiLagSplit}
\alias{LongiLagSplit.boot}
\alias{LongiLagSplit.visit}
\alias{LongiLagSplit.visit.boot}
\alias{Performance.measures}
\alias{opt.thresh}
\alias{sparseFactor}
\alias{VimPlot}
\alias{varImpPlot.RF}
\alias{createBootSamples}
\alias{multiClassSummary}
\alias{normalize}
\alias{denormalize}
\alias{getResults.ci}
\alias{F.measure.single}
\title{Utility functions: 
Some useful miscellaneous functions: plot plot.MECTree plot the 
tree part of MECTree object; heat.tree plot a sort of heat map decision trees}
\usage{
reOnly(f)

LongiLagSplit(dat, id, rhs.vars, resp.vars, order.vars = "time",
  lag = 1)

LongiLagSplit.boot(dat, id, rhs.vars, resp.vars, order.vars = "time",
  lag = 1)

LongiLagSplit.visit(dat, id, rhs.vars, resp.vars, order.vars = "time",
  visit = 1)

LongiLagSplit.visit.boot(dat, id, rhs.vars, resp.vars,
  order.vars = "time", visit = 1, parallel = TRUE)

Performance.measures(pred, obs, threshold = NULL, prevalence = NULL)

opt.thresh(pred, obs)

sparseFactor(dummy, thresh = 0.00016)

VimPlot(x, top = min(20, length(x$importance)), ...)

varImpPlot.RF(x, sort = TRUE, n.var = min(30, nrow(x$importance)),
  type = NULL, class = NULL, scale = TRUE, main = "", xlab = "",
  ...)

createBootSamples(y, times = 10, replace = TRUE, prob = NULL)

multiClassSummary(data, class.names = NULL, eps = 1e-15)

normalize(x)

denormalize(normalized, min, max)

getResults.ci(tab, alpha = 0.05)

F.measure.single(pred, labels)
}
\arguments{
\item{dat}{data.frame}

\item{id}{group id}

\item{resp.vars, rhs.vars}{response and vector of predictor names}

\item{order.vars}{variables to order longitudinal data}

\item{lag}{lag in train test split}

\item{pred, obs}{predicted probabilities and true class}

\item{threshold}{threshold}

\item{\dots}{Further arguments passed to or from other methods.}

\item{y, f, nl.n}{numeric or character vector depending on usage}

\item{cut.off, propo}{cut-off and proportions}

\item{no.last}{number of tail observations that goes into test set}

\item{allow.new.levels}{logical}

\item{formula}{formula}
}
\description{
Utility functions: 
Some useful miscellaneous functions: plot plot.MECTree plot the 
tree part of MECTree object; heat.tree plot a sort of heat map decision trees
}
\author{
Che Ngufor Ngufor.Che@mayo.edu
}
